// common/MISA.QLCH.CommonStores/src/models/vendor/IVendor.ts
export default interface IVendor {
    /**
     * Giá trị cột
     */
    CodeVendor?: string;
    NameVendor?: string;
    GroupVendor?: string;
    TelephoneVendor?: number;
    AddressVendor?: string;
    FullNameVendor?: string;
    EmailVendor?: string;
    PhoneContactVendor?: number;
    TitleVendor?: string;
    AddressItemVendor?: string;
    StatusVendor?: string;
}